refactor: 코드 리팩터링 및 패키지 구조 수정
